requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0

# Embeddings
sentence-transformers>=2.2.0
openai>=1.0.0  # If using OpenAI embeddings

# Vector Store
faiss-cpu>=1.7.0
chromadb>=0.4.0

# LLM
transformers>=4.30.0
torch>=2.0.0
langchain>=0.0.300  # Optional, for LLM abstraction

# Utils
requests>=2.25.0
tiktoken>=0.4.0
tqdm>=4.60.0
pytest>=7.0.0

# Additional dependencies for advanced features
scikit-learn>=1.0.0
pyyaml>=6.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Logging and monitoring
python-json-logger>=2.0.0

# Performance monitoring
psutil>=5.8.0 